refactor(bot): message receive user interaction action #22
