+++ bot/speeches/v1/actions/index.js
@@ -1,2 +1,2 @@
 export { default as quickReply } from './quick-reply'
-// export { default as userInteraction } from './user-interaction'
+export { default as userInteraction } from './user-interaction'
